Add spec coverage of parser configuration
Without this patch we don't have any spec coverage of the change being
made.  This patch adds two tests that makes sure the setup_puppet method
configures Puppet with the :parser setting.  We don't actually test to
make sure the configuration is read from Rspec, but the majority of the
behavior in the method is covered.